moved incrementValidators and check weight cap
