+++ x/stakeibc/keeper/validator.go
@@ -162,3 +162,114 @@ func (k Keeper) GetTotalValidatorWeight(validators []types.Validator) uint64 {
 	}
 	return totalWeight
 }
+
+// Increments the validators slash query progress tracker
+func (k Keeper) IncrementValidatorSlashQueryProgress(
+	ctx sdk.Context,
+	chainId string,
+	validatorAddress string,
+	amount sdkmath.Int,
+) error {
+	hostZone, found := k.GetHostZone(ctx, chainId)
+	if !found {
+		return types.ErrHostZoneNotFound
+	}
+
+	validator, valIndex, found := GetValidatorFromAddress(hostZone.Validators, validatorAddress)
+	if !found {
+		return types.ErrValidatorNotFound
+	}
+
+	// Increment the progress tracker
+	oldProgress := validator.SlashQueryProgressTracker
+	newProgress := validator.SlashQueryProgressTracker.Add(amount)
+	validator.SlashQueryProgressTracker = newProgress
+
+	// If the checkpoint is zero, it implies the TVL was 0 last time it was set, and we should
+	// update it here
+	// If the checkpoint is non-zero, only update it if it was just breached
+	shouldUpdateCheckpoint := true
+	if !validator.SlashQueryCheckpoint.IsZero() {
+		oldInterval := oldProgress.Quo(validator.SlashQueryCheckpoint)
+		newInterval := newProgress.Quo(validator.SlashQueryCheckpoint)
+		shouldUpdateCheckpoint = oldInterval.LT(newInterval)
+	}
+
+	// Optionally re-calculate the checkpoint
+	// Threshold of 1% means once 1% of TVL has been breached, the query is issued
+	if shouldUpdateCheckpoint {
+		validator.SlashQueryCheckpoint = k.GetUpdatedSlashQueryCheckpoint(ctx, hostZone.TotalDelegations)
+	}
+
+	hostZone.Validators[valIndex] = &validator
+	k.SetHostZone(ctx, hostZone)
+
+	return nil
+}
+
+// Increments the number of validator delegation changes in progress by 1
+// Note: This modifies the original host zone struct. The calling function must Set this host zone
+// for changes to persist
+func (k Keeper) IncrementValidatorDelegationChangesInProgress(hostZone *types.HostZone, validatorAddress string) error {
+	validator, valIndex, found := GetValidatorFromAddress(hostZone.Validators, validatorAddress)
+	if !found {
+		return errorsmod.Wrapf(types.ErrValidatorNotFound, "validator %s not found", validatorAddress)
+	}
+	validator.DelegationChangesInProgress += 1
+	hostZone.Validators[valIndex] = &validator
+	return nil
+}
+
+// Decrements the number of validator delegation changes in progress by 1
+// Note: This modifies the original host zone struct. The calling function must Set this host zone
+// for changes to persist
+func (k Keeper) DecrementValidatorDelegationChangesInProgress(hostZone *types.HostZone, validatorAddress string) error {
+	validator, valIndex, found := GetValidatorFromAddress(hostZone.Validators, validatorAddress)
+	if !found {
+		return errorsmod.Wrapf(types.ErrValidatorNotFound, "validator %s not found", validatorAddress)
+	}
+	if validator.DelegationChangesInProgress == 0 {
+		return errorsmod.Wrapf(types.ErrInvalidValidatorDelegationUpdates,
+			"cannot decrement the number of delegation updates if the validator has 0 updates in progress")
+	}
+	validator.DelegationChangesInProgress -= 1
+	hostZone.Validators[valIndex] = &validator
+	return nil
+}
+
+// Checks if any validator's portion of the weight is greater than the cap
+func (k Keeper) CheckValidatorWeightsBelowCap(ctx sdk.Context, validators []*types.Validator) error {
+	// If there's only a few validators, don't enforce this yet
+	if len(validators) < MinValidatorsBeforeWeightCapCheck {
+		return nil
+	}
+
+	// The weight cap in params is an int representing a percentage (e.g. 10 is 10%)
+	params := k.GetParams(ctx)
+	validatorWeightCap := float64(params.ValidatorWeightCap)
+
+	// Store a map of each validator weight, as well as the total
+	totalWeight := float64(0)
+	weightsByValidator := map[string]float64{}
+	for _, validator := range validators {
+		weightsByValidator[validator.Address] = float64(validator.Weight)
+		totalWeight += float64(validator.Weight)
+	}
+
+	// If the total validator weights are 0, exit prematurely
+	if totalWeight == 0 {
+		return nil
+	}
+
+	// Check if any validator exceeds the cap
+	for _, address := range utils.StringMapKeys[float64](weightsByValidator) {
+		weightPercentage := weightsByValidator[address] / totalWeight * 100
+		if weightPercentage > validatorWeightCap {
+			return errorsmod.Wrapf(types.ErrValidatorExceedsWeightCap,
+				"validator %s exceeds weight cap, has %v%% of the total weight when the cap is %v%%",
+				address, weightPercentage, validatorWeightCap)
+		}
+	}
+
+	return nil
+}
